Fast commit TODO: Update documentation file with expected deliveries and real deliveries as on today changes: 31/05/2016
Changes to put at the documentation file:
Added logic for energy formula method and modified the following classes:energyWF, accelerationWF, positionWF(Not yet finished), velocityWF and initialVHmax. Added at all the files a back button and a main button and an if statement in case of the user inputs invalid characters at te textboxes or doenst input anything.
Also added the tasks that do my classroom mate at the "swap".